Refactor more.
