--- src/cli/interfaces.ts
@@ -1,4 +1,0 @@
-// Generic error in CLI
-export interface ILoasError {
-  message: string;
-}
